Create I_GIVE_GOOD.md
Me and my charitable foundation I_GIVE_GOOD, Our ideas and deeds bring the idea of good and help to people, the community and the soul. I am studying on my own and what I have achieved is the result of a lot of work, work on myself. I ask you to be either my friends or not to prevent me from going my own way of good. I'm glad you're reading this, so it's not an accident!